requirements.txt
streamlit
google-genai
requests
bcrypt
sqlalchemy
python-dotenv